Fix update_existing_team

+++ db/migrate/20180312140952_remove_unused_column.rb
@@ -1,10 +1,10 @@
 class RemoveUnusedColumn < ActiveRecord::Migration[5.1]
   def up
-  	remove_column :teams, :overriding, if: column_exists?(:teams, :overriding)
-  	remove_column :teams, :overriding_percentage
-  	remove_column :sales, :unit_id
-  	remove_column :commissions, :position_id
-  	drop_table :units
+  	remove_column :teams, :overriding if column_exists?(:teams, :overriding)
+  	remove_column :teams, :overriding_percentage if column_exists?(:teams, :overriding_percentage)
+  	remove_column :sales, :unit_id if column_exists?(:sales, :unit_id)
+  	remove_column :commissions, :position_id if column_exists?(:commissions, :position_id)
+  	drop_table :units if table_exists?(:units)
   end
 
   def down
